# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * X Tweet Deleter API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthLoginResponse {
  url: string;
  state: string;
}

export interface PublicMetrics {
  tweet_count: number;
  followers_count: number;
  following_count: number;
  like_count?: number;
}

export interface UserInfo {
  id: string;
  name: string;
  username: string;
  profile_image_url?: string;
  public_metrics?: PublicMetrics;
}

export interface AuthCallbackResponse {
  success: boolean;
  user: UserInfo;
}

export type TweetPublicMetrics = {
  retweet_count?: number;
  like_count?: number;
  reply_count?: number;
};

export interface Tweet {
  id: string;
  text: string;
  created_at: string;
  public_metrics?: TweetPublicMetrics;
}

export interface TweetsListResponse {
  tweets: Tweet[];
  next_token?: string;
  total_fetched: number;
}

export interface DeleteTweetsRequest {
  tweet_ids: string[];
}

export interface DeleteAllTweetsRequest {
  older_than_days?: number;
  keywords?: string[];
  include_retweets?: boolean;
}

export interface DeleteTweetsResponse {
  deleted: number;
  failed: number;
  total: number;
  errors?: string[];
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export type AuthCallbackParams = {
  code: string;
  state: string;
};

export type TweetsListParams = {
  pagination_token?: string;
};