# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * X Tweet Deleter API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns the X OAuth authorization URL
 * @summary Initiate OAuth login
 */
export const AuthLoginResponse = zod.object({
  url: zod.string(),
  state: zod.string(),
});

/**
 * Handles the OAuth callback from X
 * @summary OAuth callback
 */
export const AuthCallbackQueryParams = zod.object({
  code: zod.coerce.string(),
  state: zod.coerce.string(),
});

export const AuthCallbackResponse = zod.object({
  success: zod.boolean(),
  user: zod.object({
    id: zod.string(),
    name: zod.string(),
    username: zod.string(),
    profile_image_url: zod.string().optional(),
    public_metrics: zod
      .object({
        tweet_count: zod.number(),
        followers_count: zod.number(),
        following_count: zod.number(),
        like_count: zod.number().optional(),
      })
      .optional(),
  }),
});

/**
 * Clears the session
 * @summary Logout
 */
export const AuthLogoutResponse = zod.object({
  success: zod.boolean(),
  message: zod.string().optional(),
});

/**
 * Returns the current authenticated user info
 * @summary Get current user
 */
export const AuthMeResponse = zod.object({
  id: zod.string(),
  name: zod.string(),
  username: zod.string(),
  profile_image_url: zod.string().optional(),
  public_metrics: zod
    .object({
      tweet_count: zod.number(),
      followers_count: zod.number(),
      following_count: zod.number(),
      like_count: zod.number().optional(),
    })
    .optional(),
});

/**
 * Fetch tweets for authenticated user
 * @summary List tweets
 */
export const TweetsListQueryParams = zod.object({
  pagination_token: zod.coerce.string().optional(),
});

export const TweetsListResponse = zod.object({
  tweets: zod.array(
    zod.object({
      id: zod.string(),
      text: zod.string(),
      created_at: zod.string(),
      public_metrics: zod
        .object({
          retweet_count: zod.number().optional(),
          like_count: zod.number().optional(),
          reply_count: zod.number().optional(),
        })
        .optional(),
    }),
  ),
  next_token: zod.string().optional(),
  total_fetched: zod.number(),
});

/**
 * Delete an array of tweet IDs
 * @summary Delete specific tweets
 */
export const TweetsDeleteBody = zod.object({
  tweet_ids: zod.array(zod.string()),
});

export const TweetsDeleteResponse = zod.object({
  deleted: zod.number(),
  failed: zod.number(),
  total: zod.number(),
  errors: zod.array(zod.string()).optional(),
});

/**
 * Delete all tweets with optional filters
 * @summary Delete all tweets
 */
export const TweetsDeleteAllBody = zod.object({
  older_than_days: zod.number().optional(),
  keywords: zod.array(zod.string()).optional(),
  include_retweets: zod.boolean().optional(),
});

export const TweetsDeleteAllResponse = zod.object({
  deleted: zod.number(),
  failed: zod.number(),
  total: zod.number(),
  errors: zod.array(zod.string()).optional(),
});

/**
 * Unlike all liked tweets
 * @summary Delete all likes
 */
export const TweetsDeleteLikesResponse = zod.object({
  deleted: zod.number(),
  failed: zod.number(),
  total: zod.number(),
  errors: zod.array(zod.string()).optional(),
});