# lib/api-zod/src/generated/types/authCallbackResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * X Tweet Deleter API
 * OpenAPI spec version: 0.1.0
 */
import type { UserInfo } from "./userInfo";

export interface AuthCallbackResponse {
  success: boolean;
  user: UserInfo;
}