# lib/api-zod/src/generated/types/deleteAllTweetsRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * X Tweet Deleter API
 * OpenAPI spec version: 0.1.0
 */

export interface DeleteAllTweetsRequest {
  older_than_days?: number;
  keywords?: string[];
  include_retweets?: boolean;
}