# lib/api-zod/src/generated/types/deleteTweetsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * X Tweet Deleter API
 * OpenAPI spec version: 0.1.0
 */

export interface DeleteTweetsResponse {
  deleted: number;
  failed: number;
  total: number;
  errors?: string[];
}