# lib/api-zod/src/generated/types/publicMetrics.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * X Tweet Deleter API
 * OpenAPI spec version: 0.1.0
 */

export interface PublicMetrics {
  tweet_count: number;
  followers_count: number;
  following_count: number;
  like_count?: number;
}