# lib/api-zod/src/generated/types/tweet.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * X Tweet Deleter API
 * OpenAPI spec version: 0.1.0
 */
import type { TweetPublicMetrics } from "./tweetPublicMetrics";

export interface Tweet {
  id: string;
  text: string;
  created_at: string;
  public_metrics?: TweetPublicMetrics;
}