# lib/api-zod/src/generated/types/tweetsListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * X Tweet Deleter API
 * OpenAPI spec version: 0.1.0
 */
import type { Tweet } from "./tweet";

export interface TweetsListResponse {
  tweets: Tweet[];
  next_token?: string;
  total_fetched: number;
}