# lib/api-zod/src/generated/types/userInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * X Tweet Deleter API
 * OpenAPI spec version: 0.1.0
 */
import type { PublicMetrics } from "./publicMetrics";

export interface UserInfo {
  id: string;
  name: string;
  username: string;
  profile_image_url?: string;
  public_metrics?: PublicMetrics;
}